ArtifactWrappers.jl

A lightweight wrapper around Julia Artifacts.

ArtifactWrappers.ArtifactFile — Type

ArtifactFile

A single data file to be downloaded, containing both the url and the name to use locally.

Fields

url URL pointing to data to be downloaded

filename Local name used for downloaded online data

ArtifactWrappers.get_data_folder — Function

get_data_folder(art_wrap::ArtifactWrapper)

Get local folder of dataset defined in art_wrap.

Example:

dataset_path = get_data_folder(dataset)

ArtifactWrappers.ArtifactWrapper — Type

ArtifactWrapper

A set of data files to be downloaded, grouped by data_name. Example:

dataset = ArtifactWrapper(
    @__DIR__,
    isempty(get(ENV, "CI", "")),
    "MyDataSet",
    ArtifactFile[
        ArtifactFile(
            url="https://..../SomeNetCDF1.nc",
            filename="experiment1.nc",
        ),
        ArtifactFile(
            url="https://..../SomeNetCDF2.nc",
            filename="experiment2.nc",
        ),
    ]
)

Fields

artifact_dir Directory to store artifact / data

local_run Locally running (not using CI)

artifact_toml Path to the used Artifacts.toml

data_name Unique name of dataset

artifact_files Array of ArtifactFile's, grouped by this dataset